suggested changes

+++ flow/connectors/bigquery/qrep_avro_sync.go
@@ -284,8 +284,8 @@ func DefineAvroSchema(dstTableName string,
 }
 
 func GetAvroType(bqField *bigquery.FieldSchema) (interface{}, error) {
-	avroNumericPrecision := bqField.Precision
-	avroNumericScale := bqField.Scale
+	avroNumericPrecision := int16(bqField.Precision)
+	avroNumericScale := int16(bqField.Scale)
 	if avroNumericPrecision > 38 || avroNumericPrecision <= 0 ||
 		avroNumericScale > 9 || avroNumericScale <= 0 {
 		avroNumericPrecision = numeric.PeerDBNumericPrecision
