+++ flow/model/numeric/scale.go
@@ -15,9 +15,11 @@ func StripTrailingZeros(value string) string {
 	return value
 }
 
-func ParseNumericTypmod(typmod int32) (int, int) {
+// This is to reverse what make_numeric_typmod of Postgres does:
+// https://github.com/postgres/postgres/blob/21912e3c0262e2cfe64856e028799d6927862563/src/backend/utils/adt/numeric.c#L897
+func ParseNumericTypmod(typmod int32) (int16, int16) {
 	offsetMod := typmod - 4
-	precision := int((offsetMod >> 16) & 0x7FFF)
-	scale := int(offsetMod & 0x7FFF)
+	precision := int16((offsetMod >> 16) & 0x7FFF)
+	scale := int16(offsetMod & 0x7FFF)
 	return precision, scale
 }
